requirements.txt
python-dotenv==1.0.0
websockets==11.0.3
openai==1.3.0
numpy==1.24.3
pydub==0.25.1
scipy==1.10.1
